+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 * feat: Add database Ca Certificate download URL by @aurelien-reeves-scalingo in https://github.com/Scalingo/scalingo.js/pull/541
 
+## 0.17.17
+
+*feat: add apiPing, apiUpgrade, and apiAction methods to Databases class with corresponding tests
+
+## 0.17.16
+
+* feat: add apiUpdate method to update database configuration via dbaas API
+* feat: add API methods for database plan, metrics, health, instances status, lag, logs, and logs archives
+
 ## 0.17.15
 
 * chore(deps-dev): bump the development group with 7 updates by @dependabot[bot] in https://github.com/Scalingo/scalingo.js/pull/532